Moved specs with file IO to io_spec
