+++ package.json
@@ -14,12 +14,12 @@
     "afinn-165": "^1.0.2",
     "apparatus": "^0.0.10",
     "json-stable-stringify": "^1.0.1",
-    "pegjs": "^0.10.0",
     "phantomjs": "^2.1.7",
     "puppeteer": "^1.11.0",
     "sylvester": "^0.0.12",
     "underscore": "^1.3.1",
-    "webpack": "^4.28.2"
+    "webpack": "^4.28.2",
+    "wordnet-db": "^3.1.11"
   },
   "devDependencies": {
     "browserfs": "^1.4.3",
@@ -29,6 +29,7 @@
     "gulp-jasmine-phantom": "^3.0.0",
     "jasmine": "^3.3.1",
     "jasmine-core": "^3.3.0",
+    "pegjs": "^0.10.0",
     "proxyquire": "^1.8.0",
     "rimraf": "^2.6.2",
     "sinon": "^1.12.2",
@@ -39,7 +40,7 @@
     "benchmark": "node benchmarks",
     "clean": "rimraf *~ #* *classifier.json",
     "test": "NODE_PATH=. node ./node_modules/jasmine/bin/jasmine.js --random=false spec/*_spec.js",
-    "test_io": "NODE_PATH=. jasmine ./io_spec/*.js"
+    "test_io": "NODE_PATH=. node ./node_modules/jasmine/bin/jasmine.js --random=false io_spec/*_spec.js"
   },
   "license": "MIT",
   "author": "Chris Umbel <chris@chrisumbel.com>",
