--- spec/initBrowserFS.js
@@ -1,3 +1,0 @@
-
-var fs = require('browserfs');
-fs.initialise
