+++ spec/tfidf_spec.js
@@ -217,58 +217,6 @@ describe('tfidf', function() {
             expect( tfidf.tfidf("node", 0) ).toBe( 1 * ( 1 + Math.log( 4.0 / 4.0 ) ));
         });
 
-        /*
-        // Test tf-idf computation on files loaded using readFileSync
-        it("should load documents from files", function(){
-
-            tfidf = new TfIdf();
-
-            tfidf.addFileSync("spec/test_data/tfidf_document1.txt", null, {node: 0, ruby:1});
-            tfidf.addFileSync("spec/test_data/tfidf_document2.txt", null, {node:1, ruby:3});
-            tfidf.addFileSync("spec/test_data/tfidf_document3.txt", null, {node:0, ruby:3});
-            tfidf.addFileSync("spec/test_data/tfidf_document4.txt", null, {node:2, ruby:1});
-
-            var correctCalculations = [
-                1 * (1 + Math.log( 4.0 / 4.0 )),
-                0,
-                2 * (1 + Math.log( 4.0 / 4.0 )),
-                1 * (1 + Math.log( 4.0 / 3.0 ))
-            ];
-
-            tfidf.tfidfs('node', function(i, measure, k) {
-                expect(measure).toBe(correctCalculations[k.node]);
-            });
-
-            tfidf.tfidfs('ruby', function(i, measure, k) {
-                expect(measure).toBe(correctCalculations[k.ruby]);
-            });
-
-            // addFileSync with restoreCache flag set to true.
-            tfidf.addFileSync("spec/test_data/tfidf_document4.txt", null, {node:2, ruby:1}, true);
-        });
-        */
-
-        /*
-        // Test idf caching when adding documents from addFileSync
-        it("should update a terms tf-idf score after adding documents from addFileSync", function(){
-            tfidf = new TfIdf();
-
-            // Add 2 documents
-            tfidf.addFileSync("spec/test_data/tfidf_document1.txt", null, 0);
-            tfidf.addFileSync("spec/test_data/tfidf_document2.txt", null, 1);
-
-            // check the tf-idf for 'node'
-            expect( tfidf.tfidf("node", 0) ).toBe( 1 * (1 + Math.log( 2.0 / 2.0 ) ));
-
-            // Add 2 more documents
-            tfidf.addFileSync("spec/test_data/tfidf_document3.txt");
-            tfidf.addFileSync("spec/test_data/tfidf_document4.txt");
-
-            // Ensure that the tf-idf in the same document has changed to reflect the new idf.
-            expect( tfidf.tfidf("node", 0) ).toBe( 1 * (1 + Math.log( 4.0 / 4.0 ) ));
-        });
-        */
-
         // Test idf.setTokenizer
         it('should allow for specific types of Tokenizers', function(){
             tfidf = new TfIdf();
@@ -305,25 +253,5 @@ describe('tfidf', function() {
             expect( function() { tfidf.setTokenizer({}); } ).toThrow(new Error('Expected a valid Tokenizer'));
         });
 
-        /*
-        // Test encoding for addFileSync
-        it('should use the specified encoding for addFileSync', function(){
-
-            tfidf = new TfIdf();
-
-            tfidf.addFileSync('spec/test_data/tfidf_document1.txt', 'base64');
-            tfidf.addFileSync('spec/test_data/tfidf_document1.txt', 'utf8');
-
-            expect( tfidf.tfidf('dghpcybkb2n1bwvudcbpcybhym91dcbub2rllg', 0) ).toBe( 1 * ( 1 + Math.log( 2.0 / 2.0 ) ));
-        });
-        */
-
-        // Test encoding check for addFileSync
-        it('should require a valid encoding for addFileSync', function(){
-
-            tfidf = new TfIdf();
-
-            expect( function() { tfidf.addFileSync('spec/test_data/tfidf_document1.txt', 'foobar'); } ).toThrow(new Error('Invalid encoding: foobar'));
-        });
     });
 });
